+++ jjtemplate-compiler/src/main/java/io/github/sibmaks/jjtemplate/compiler/runtime/ExpressionSourceRenderer.java
@@ -95,7 +95,7 @@ public final class ExpressionSourceRenderer implements ExpressionVisitor<String>
 
     @Override
     public String visitRange(RangeExpression expr) {
-        return render(expr.name) + " range " + expr.itemVariableName + "," + expr.indexVariableName + " of "
+        return render(expr.name) + " range " + expr.firstVariableName + "," + expr.secondVariableName + " of "
                 + render(expr.source);
     }